+++ client/src/components/dialog/EditAppDialog.js
@@ -1,6 +1,5 @@
-import PropTypes from 'prop-types'
-import React, { Component } from 'react'
-import TextField from 'material-ui/TextField'
+// eslint-disable-next-line react/no-deprecated
+import React, { PropTypes, Component } from 'react'
 import { connect } from 'react-redux'
 import DialogBase from './DialogBase'
 import { editApp } from '../../actions/actionCreators'
@@ -26,10 +25,6 @@ export class EditAppDialog extends Component {
     }
 
     render() {
-        const fieldStyle = {
-            display: 'block',
-            width: '100%',
-        }
         const app = this.props.app
         return (
             <DialogBase
@@ -41,6 +36,7 @@ export class EditAppDialog extends Component {
                         appName: app.name,
                         description: app.description,
                         appType: app.appType,
+                        sourceUrl: app.sourceUrl,
                         developerName: app.developer.name,
                         developerEmail: app.developer.email,
                         developerAddress: app.developer.address,
@@ -58,8 +54,7 @@ export class EditAppDialog extends Component {
 
 EditAppDialog.propTypes = {
     app: PropTypes.object,
-    appId: PropTypes.string,
-    addVersion: PropTypes.func,
+    editApp: PropTypes.func.required,
 }
 
 const mapDispatchToProps = dispatch => ({